Replaced signed shorts with int_fast16_t, unsigned shorts with uint_fast16_t, chars with int_fast8_t, and unsigned chars with uint_fast8_t Made enums more "strongly typed" Replaced most ternary operators either with if statements or switches Simplified code where possible
To do:
	Make macros for conversions
	**TEST**

+++ PROS/Prototype/include/kalmanFilter.h
@@ -1,12 +1,14 @@
 #ifndef KALMANFILTER_H_
 #define KALMANFILTER_H_
 
+#include <API.h>
+
 float filterDrive[6];
 float filterDriveL[6], filterDriveR[6];
 float filterArmL[6], filterArmR[6];
 //Remember to initialize arrays in init.c
 
-int getSensor(float values[], int sensorValue);
+uint_fast16_t getSensor(float values[], uint_fast16_t sensorValue);
 /*
 values[] array format:
 KG(0), estimate(1), previousEstimate(2), errorEstimate(3), previousErrorEstimate(4), errorMeasurement(5)
